move generic_mocks to a specific file
