+++ cypress/support/datagouv_mocks.js
@@ -48,32 +48,6 @@ Cypress.Commands.add(
   }
 )
 
-const allowedDomains = [
-  'localhost',
-  '127.0.0.1',
-  'content-autofill.googleapis.com',
-  'android.clients.google.com'
-]
-
-Cypress.Commands.add('catchUnmockedRequests', () => {
-  const escapedDomains = allowedDomains
-    .map((domain) => domain.replace(/\./g, '\\.'))
-    .join('|')
-  const regex = new RegExp(`^https?:\\/\\/(?!(${escapedDomains})(:|\/|$)).*`)
-
-  cy.intercept(regex, (req) => {
-    throw new Error(
-      `Unmocked external API call detected: ${req.method} ${req.url}`
-    )
-  }).as('catchExternalRequests')
-})
-
-Cypress.Commands.add('allowExternalRequests', () => {
-  cy.intercept('**', (req) => {
-    req.continue()
-  }).as('catchExternalRequests')
-})
-
 Cypress.Commands.add('mockGristImages', () => {
   cy.readFile('public/blank_state/file.svg').then((svgContent) => {
     cy.intercept(
